Reogranise
Separate free functions into game and gameOver folders.
