+++ include/gameOver.hpp
@@ -1,0 +1,9 @@
+#ifndef END_GAME_HPP
+#define END_GAME_HPP
+
+#include "./wrappers/iRenderWindow.hpp"
+#include "./wrappers/iText.hpp"
+
+void displayGameOverScreen(IRenderWindow &window, IText &gameOverText, IText &playAgainText);
+
+#endif
