--- src/freeFunctions/endGame.cpp
@@ -1,5 +1,0 @@
-void endGame(bool &isPlaying, bool &gameOver)
-{
-  isPlaying = false;
-  gameOver = true;
-}
