--- src/freeFunctions/fireLaser.cpp
@@ -1,6 +1,0 @@
-#include "../../include/models/iLaserCannon.hpp"
-
-void fireLaser(ILaserCannon &cannon)
-{
-  cannon.fire();
-}
