+++ src/freeFunctions/game/endGame.cpp
@@ -1,0 +1,5 @@
+void endGame(bool &isPlaying, bool &gameOver)
+{
+  isPlaying = false;
+  gameOver = true;
+}
