+++ src/freeFunctions/game/fireLaser.cpp
@@ -1,0 +1,6 @@
+#include "../../../include/models/iLaserCannon.hpp"
+
+void fireLaser(ILaserCannon &cannon)
+{
+  cannon.fire();
+}
