+++ src/freeFunctions/game/moveLaser.cpp
@@ -1,0 +1,6 @@
+#include "../../../include/models/iLaser.hpp"
+
+void moveLaser(ILaser &laser)
+{
+  laser.move();
+}
