+++ src/freeFunctions/game/moveLaserCannon.cpp
@@ -1,0 +1,6 @@
+#include "../../../include/models/iLaserCannon.hpp"
+
+void moveLaserCannon(ILaserCannon &cannon, float x)
+{
+  cannon.move(x);
+}
