--- src/freeFunctions/moveLaser.cpp
@@ -1,6 +1,0 @@
-#include "../../include/models/iLaser.hpp"
-
-void moveLaser(ILaser &laser)
-{
-  laser.move();
-}
