--- src/freeFunctions/moveLaserCannon.cpp
@@ -1,6 +1,0 @@
-#include "../../include/models/iLaserCannon.hpp"
-
-void moveLaserCannon(ILaserCannon &cannon, float x)
-{
-  cannon.move(x);
-}
