--- tests/freeFunctions/endGame.test.cpp
@@ -1,20 +1,0 @@
-#include <gtest/gtest.h>
-#include "../../include/spaceInvaders.hpp"
-
-TEST(endGame, changesIsPlayingToFalse)
-{
-  bool isPlaying = true;
-  bool gameOver = false;
-
-  endGame(isPlaying, gameOver);
-  EXPECT_EQ(isPlaying, false);
-}
-
-TEST(endGame, changesGameOverToTrue)
-{
-  bool isPlaying = true;
-  bool gameOver = false;
-
-  endGame(isPlaying, gameOver);
-  EXPECT_EQ(gameOver, true);
-}
