+++ tests/freeFunctions/game/fireLaser.test.cpp
@@ -1,0 +1,11 @@
+#include "../../../include/game.hpp"
+#include "../../mockModels/mockLaserCannon.hpp"
+
+TEST(fireLaser, callsFireOnLaserCannon)
+{
+  MockLaserCannon cannon;
+
+  EXPECT_CALL(cannon, fire())
+      .Times(1);
+  fireLaser(cannon);
+}
