+++ tests/freeFunctions/game/moveLaser.test.cpp
@@ -1,0 +1,11 @@
+#include "../../../include/game.hpp"
+#include "../../mockModels/mockLaser.hpp"
+
+TEST(moveLaser, callsMoveOnLaser)
+{
+  MockLaser laser;
+
+  EXPECT_CALL(laser, move())
+      .Times(1);
+  moveLaser(laser);
+}
