+++ tests/freeFunctions/game/moveLaserCannon.test.cpp
@@ -1,0 +1,11 @@
+#include "../../../include/game.hpp"
+#include "../../mockModels/mockLaserCannon.hpp"
+
+TEST(moveLaserCannon, callsMoveOnTheLaserCannon)
+{
+  MockLaserCannon cannon;
+
+  EXPECT_CALL(cannon, move(150))
+      .Times(1);
+  moveLaserCannon(cannon, 150);
+}
