--- tests/freeFunctions/moveLaser.test.cpp
@@ -1,11 +1,0 @@
-#include "../../include/spaceInvaders.hpp"
-#include "../mockModels/mockLaser.hpp"
-
-TEST(moveLaser, callsMoveOnLaser)
-{
-  MockLaser laser;
-
-  EXPECT_CALL(laser, move())
-      .Times(1);
-  moveLaser(laser);
-}
